Abstraction for vlm support
